refactor: api 응답 타입 types/apis 폴더로 이동
