+++ src/pages/onboarding/types/apis/step1.types.ts
@@ -1,0 +1,14 @@
+export interface FloorPlanList {
+  id: number;
+  form: string;
+  structure: string;
+  floorPlanImage: string;
+}
+
+export interface FloorPlanResponse {
+  code: number;
+  msg: string;
+  data: {
+    floorPlanList: FloorPlanList[];
+  };
+}
